Corrected turning for 'd', added docs
