--- SerialMonitor/install_dep.sh
@@ -1,6 +1,0 @@
-sudo apt-get install -y python3 libatlas-base-dev python3-numpy python3-pandas python3-serial python3-matplotlib 
-
-echo "export Path=/home/pi/.local/bin:$Path" >> ~/.bashrc
-. ~/.bashrc
-
-#python3 -m pip install pandas pyserial matplotlib
