--- SerialMonitor/sys_id_template.csv
@@ -1,3 +1,0 @@
-cf, Q, 0.002
-chhf, P, 50, 50, 0.5
-cf, Q, -1
